NextjsAnswer cleanup (remove redundant props, move/rename file to follow base Answer naming)
