--- packages/create-sitecore-jss/src/initializers/models.ts
@@ -1,6 +1,0 @@
-import { Answer } from '../common/Answer';
-export interface NextjsAnswer extends Answer {
-  prerender: string;
-  hostName: string;
-  appPrefix?: boolean;
-}
