+++ packages/create-sitecore-jss/src/initializers/nextjs/NextjsAnswer.ts
@@ -1,0 +1,5 @@
+import { Answer } from '../../common/Answer';
+
+export interface NextjsAnswer extends Answer {
+  prerender: string;
+}
